Fixed images deletion and added new folder for each entity images

+++ GraduationProject_BL/Managers/BannerManager.cs
@@ -18,7 +18,7 @@ namespace GraduationProject_BL.Managers
     {
         private readonly IRepository<Banners> repository;
         private readonly ITranslations<BannerTranslation> translations;
-        private readonly string retrievePath = "assets\\images";
+        private readonly string retrievePath = "assets\\images\\banners";
 
 
         public BannerManager(IRepository<Banners> _repository, ITranslations<BannerTranslation> _translations)
@@ -32,8 +32,17 @@ namespace GraduationProject_BL.Managers
         public async Task DeleteAsync(int id)
         {
 
-            await translations.DeleteAsync(id);
-            await repository.DeleteAsync(id);
+            var banners = await repository.GetAllAsync();
+            var banner = banners.Find(d => d.Id == id);
+            if (banner != null)
+            {
+                if (banner.ImagePath != null)
+                {
+                    DeleteBannerImage(banner.ImagePath);
+                }
+                await translations.DeleteAsync(id);
+                await repository.DeleteAsync(id);
+            }
         }
 
         public async Task<List<BannerDTO>> GetAllAsync(string lang)
